Refined Error handling
Updated error handling to be more easier to read. Instead of using try/catch everywhere and using .then/.catch, I've added a centralized middleware app where all errors happened by the app to automatically send error status code and stopped it from crashing the server. I also added a place where specific common errors can sent to the front-end with much more details on what the user has made to create the error. I've added prowrap as to help make other specfic errors to be redirected in the app itself if there's a certain control flow I wanted to take when the app encounters that error. An example would be an error for the settings.json file not existing. The app now must create that file with brand new settings and send that to the user. This is much better than sending a error message and having to let the user click the button again to make it work.